requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.2
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-appconf==1.0.4
django-imagekit==4.0.2
django-import-export==2.0.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.9
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jdcal==1.4.1
MarkupPy==1.14
mccabe==0.6.1
msgpack==0.6.2
numpy==1.19.2
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.5
pdfkit==0.6.1
pilkit==2.0
Pillow==7.1.1
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.50
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tablib==1.1.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
websocket-client==0.57.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.1.2
